backend/fastapi_app/requirements.txt
fastapi[standard]
langchain-google-genai

python-dotenv
